+++ src/lex/split_shell.c
@@ -6,39 +6,11 @@
 /*   By: mely-pan <mely-pan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/15 12:57:06 by mely-pan          #+#    #+#             */
-/*   Updated: 2025/04/13 17:25:39 by htrindad         ###   ########.fr       */
+/*   Updated: 2025/04/13 18:04:03 by htrindad         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../minishell.h"
-
-static size_t	ft_count_words(char const *s, t_ms *ms)
-{
-	size_t	i;
-	size_t	count;
-	char	*tmp;
-
-	i = 0; //len
-	count = 0; // space
-	while (s[i])
-	{
-		while (s[i] == ' ')
-			i++;
-		if (!s[i])
-			break ;
-		count++;
-		tmp = ft_substr(s, i, 2);
-		if (is_special_char(s[i]) || s[i] == '\'' || s[i] == '\"')
-			i += iteration_cases(s, i, ms->scases, ms);
-		else
-		{
-			while (s[i] && s[i] != ' ' && !is_special_char(s[i]))
-				i += iterate_through_q(s, i, ms);
-		}
-		free(tmp);
-	}
-	return (count);
-}
 
 static bool	ft_safe_allocate(char **array, int index, size_t len)
 {
@@ -77,16 +49,19 @@ static bool	ft_filling_arr(char **array, char const *s, t_ms *ms)
 	return (false);
 }
 
-char	**ft_split_shell(t_ms *shell)
+char	***ft_split_shell(t_ms *shell)
 {
 	size_t	words;
-	char	**array;
+	size_t	cases;
+	char	***array;
 	char	*new_s;
 
 	if (!shell->input)
 		return (NULL);
 	words = ft_count_words(shell->input, shell);
-	array = ft_calloc(words + 1, sizeof(char *));
+	cases = count_cases(shell->input, shell);
+	op_funcs(array, words, count);
+	//array = ft_calloc(words + 1, sizeof(char *)); // Gonna operate on this function
 	if (!array)
 		return (NULL);
 	if (has_env_var(shell->input))
