+++ src/lex/split_shell_utils.c
@@ -6,14 +6,14 @@
 /*   By: mely-pan <mely-pan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/15 12:57:01 by mely-pan          #+#    #+#             */
-/*   Updated: 2025/04/13 17:25:54 by htrindad         ###   ########.fr       */
+/*   Updated: 2025/04/13 17:52:54 by htrindad         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../minishell.h"
 
 // spec_case will compare the sub with one of the cases "< > | << >>"
-bool	spec_case(char *sub, char **cases, size_t *l)
+size_t	spec_case(char *sub, char **cases, size_t *l)
 {
 	int	i;
 
@@ -24,10 +24,10 @@ bool	spec_case(char *sub, char **cases, size_t *l)
 		if (!ft_strncmp(sub, cases[i], ft_strlen(cases[i])))
 		{
 			*l = ft_strlen(cases[i]);
-			return (true);
+			return (*l);
 		}
 	}
-	return (false);
+	return (0);
 }
 
 // iteration cases return the length of the next token.
